updated the migrator, data models and data context to allow for cards that act as multiple cards i.e. cards that share a Multiverse ID such as "Alive" & "Well" multiverseID: 369041
there is some cleanup required for the MagicMigrator. the code that handles sharing multiverseID will work for cards that dont share so I need to remove the old way of parsing nonsharing cards so that they go through the same code as the other cards. basically doing this will make it so that I dont have methods such as getType and getTypeDual as getTypeDual can handle cards that would normally be handled by getType but the reverse is not true